fix: specify KES sig size and restrict tx inputs to 16 bits

+++ crates/amaru-consensus/src/stages/mempool/test_setup.rs
@@ -123,8 +123,8 @@ pub fn te_record_metrics(at_stage: &str, metrics: MempoolMetrics) -> TraceEntry 
     ))
 }
 
-pub fn create_transaction(input_index: usize) -> WithOriginalBytes<Transaction> {
-    let tx_input = TransactionInput { transaction_id: Hash::new([1; TRANSACTION_BODY]), index: input_index as u64 };
+pub fn create_transaction(input_index: u16) -> WithOriginalBytes<Transaction> {
+    let tx_input = TransactionInput { transaction_id: Hash::new([1; TRANSACTION_BODY]), index: input_index };
     let body = TransactionBody::new([tx_input], [], 0);
     Transaction {
         body,
